--- a/csrc/CMakeLists.txt
+++ b/csrc/CMakeLists.txt
@@ -12,3 +12,4 @@ endif()
 if (BUILD_TORCH_PLUGIN)
     add_subdirectory(torch)
 endif (BUILD_TORCH_PLUGIN)
+