--- a/csrc/engine/ascend_direct_transport/CMakeLists.txt
+++ b/csrc/engine/ascend_direct_transport/CMakeLists.txt
@@ -1,0 +1,4 @@
+file(GLOB ASCEND_SOURCES "*.cpp")
+
+add_library(ascend_direct OBJECT ${ASCEND_SOURCES})
+target_link_libraries(ascend_direct PUBLIC _slime_engine)