--- a/csrc/engine/ascend_direct_transport/ascend_direct_transport.h
+++ b/csrc/engine/ascend_direct_transport/ascend_direct_transport.h
@@ -1,0 +1,38 @@
+#pragma once
+
+#include "adxl/adxl_engine.h"
+#include "engine/assignment.h"
+#include <map>
+#include <memory>
+#include <string>
+#include <unordered_set>
+
+namespace slime {
+
+class AscendDirectContext {
+public:
+    AscendDirectContext() = default;
+    ~AscendDirectContext();
+
+    int init(const std::string& host, int port);
+
+    int read_batch(AssignmentBatch& batch, const std::string& host, int port);
+
+    int register_memory_region(const std::string& location, uintptr_t addr, size_t length);
+
+    int unregister_memory_region(uintptr_t addr);
+
+    int connect(const std::string& host, int port);
+
+    int disconnect(const std::string& adxl_engine_name);
+
+    int disconnect(const std::string& host, int port);
+
+private:
+    const static int                     CONNECT_TIMEOUT_MILLIS = 60 * 1000;
+    std::unique_ptr<adxl::AdxlEngine>    adxl_                  = nullptr;
+    std::map<uintptr_t, adxl::MemHandle> addr_to_memhandle_;
+    std::unordered_set<std::string>      connected_engines_;
+};
+
+}  // namespace slime