--- a/tests/cpp/ascend_direct/ascend_direct_perf.cpp
+++ b/tests/cpp/ascend_direct/ascend_direct_perf.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+
+#include "engine/ascend_direct_transport/ascend_direct_transport.h"
+
+int main(int argc, char** argv)
+{
+    std::cout << "hello world" << std::endl;
+    slime::AscendDirectContext ascend_direct;
+}