This PR adds Ascend Direct Transfer and Perf Test. Since Huawei's ADXL is not opened to public yet, we put a mock ADXL class in this PR. 

Performance Test Usage:

1. Start Huawei docker container on a Huawei machine
2. `cd` to repo root
3.  Compile with following commands:
```
mkdir build
cd build
cmake -DBUILD_TEST=ON -DBUILD_ASCEND_DIRECT=ON ..
make -j
```
4. Run "initiator" and "target" two instances, for example:
```
./bin/ascend_direct_perf --mode=target --localhost="100.97.164.197" --local_port=16789 --remote_host="100.97.164.197" --remote_port=16777
./bin/ascend_direct_perf --mode=initiator --localhost="100.97.164.197" --local_port=16777 --remote_host="100.97.164.197" --remote_port=16789
```

5. You could see perf result like these from initiator output:
```
Initializing ACL
Running initiator or target
running initiator on 100.97.164.197:16777
receiving addr info through zmq from 100.97.164.197:16789
waiting for initiator do ascend data transfer perf
Block iteration 0 test completed: duration 26us, block size 32KB, total size 1024KB, throughput 40.33 GB/s
Block iteration 1 test completed: duration 5us, block size 64KB, total size 2048KB, throughput 419.43 GB/s
Block iteration 2 test completed: duration 3us, block size 128KB, total size 4096KB, throughput 1398.10 GB/s
Block iteration 3 test completed: duration 4us, block size 256KB, total size 8192KB, throughput 2097.15 GB/s
Block iteration 4 test completed: duration 3us, block size 512KB, total size 16384KB, throughput 5592.41 GB/s
Block iteration 5 test completed: duration 3us, block size 1024KB, total size 32768KB, throughput 11184.81 GB/s
Block iteration 6 test completed: duration 3us, block size 2048KB, total size 65536KB, throughput 22369.62 GB/s
Block iteration 7 test completed: duration 3us, block size 4096KB, total size 131072KB, throughput 44739.24 GB/s
Block iteration 8 test completed: duration 4us, block size 8192KB, total size 262144KB, throughput 67108.86 GB/s
Block iteration 9 test completed: duration 4us, block size 16384KB, total size 524288KB, throughput 134217.73 GB/s
ending ascend transfer perf, releasing resources
```